rename files to *state
