+++ server/source/serverlib/game/gamelogic.cpp
@@ -1,8 +1,8 @@
 #include "gamelogic.h"
 
 #include <game/game.h>
-#include <game/stages/endstage.h>
-#include <game/stages/runstage.h>
+#include <game/turn/endstate.h>
+#include <game/turn/startstate.h>
 #include <game/turn/freestate.h>
 #include <game/turn/jailstate.h>
 #include <game/turn/rollstate.h>
