+++ server/source/serverlib/game/turn/endstate.h
