+++ server/source/serverlib/game/turn/jailstate.cpp
@@ -1,6 +1,6 @@
 #include "jailstate.h"
 
-#include <game/stages/runstage.h>
+#include <game/turn/startstate.h>
 #include <game/turn/freestate.h>
