+++ server/tests/server-test/CMakeLists.txt
@@ -24,7 +24,7 @@ set(sources
 	tests/gameconnection_test.cpp
     tests/request_test.cpp
     tests/connection_test.cpp
-    tests/runstage_test.cpp
+    tests/startstage_test.cpp
     tests/websocketserver_test.cpp
     test_utils/mocked_database.cpp
     test_utils/mocked_database.h
